+++ drivers/crypto/crypto_stm32.c
@@ -8,7 +8,7 @@
 #include <kernel.h>
 #include <device.h>
 #include <sys/__assert.h>
-#include <crypto/cipher.h>
+#include <crypto/crypto.h>
 #include <drivers/clock_control/stm32_clock_control.h>
 #include <drivers/clock_control.h>
 #include <sys/byteorder.h>
