+++ drivers/crypto/crypto_tc_shim.c
@@ -14,7 +14,7 @@
 #include <tinycrypt/constants.h>
 #include <tinycrypt/utils.h>
 #include <string.h>
-#include <crypto/cipher.h>
+#include <crypto/crypto.h>
 #include "crypto_tc_shim_priv.h"
 
 #define LOG_LEVEL CONFIG_CRYPTO_LOG_LEVEL
