+++ drivers/ieee802154/ieee802154_cc2520.c
@@ -33,7 +33,7 @@ LOG_MODULE_REGISTER(LOG_MODULE_NAME);
 
 #ifdef CONFIG_IEEE802154_CC2520_CRYPTO
 
-#include <crypto/cipher.h>
+#include <crypto/crypto.h>
 #include <crypto/cipher_structs.h>
 
 #endif /* CONFIG_IEEE802154_CC2520_CRYPTO */
